fix bug

+++ chain/src/engine/config.rs
@@ -7,7 +7,7 @@
 pub struct EngineConf {
     pub max_block_txs: usize,
     pub max_block_size: usize,
-    pub chain_id: u64, // sub chain id
+    pub chain_id: u32, // sub chain id
     pub unstable_block: u64, // The number of blocks that are likely to fall back from the fork
     pub fast_sync: bool,
     pub data_dir: String,
@@ -78,7 +78,7 @@ impl EngineConf {
         cnf.fast_sync = ini_must_bool(sec, "fast_sync", false);
 
         let sec_mint = &ini_section(ini, "mint");
-        cnf.chain_id = ini_must_u64(sec_mint, "chain_id", 0);
+        cnf.chain_id = ini_must_u64(sec_mint, "chain_id", 0) as u32;
 
         // HAC miner
         let sec_miner = &ini_section(ini, "miner");
